债权价值计算

+++ etl-base/etl-base-common/src/main/java/com/etl/base/common/util/DateUtils.java
@@ -109,8 +109,8 @@ public class DateUtils {
    * @param end
    * @return
    */
-  public static long intervalDay(long start, long end){
-    return ChronoUnit.DAYS.between(toLocalDateTime(start), toLocalDateTime(end));
+  public static int intervalDay(long start, long end){
+    return (int)ChronoUnit.DAYS.between(toLocalDateTime(start), toLocalDateTime(end));
   }
 
   /**
